requirements.txt
# 🔹 Core LangChain
langchain>=0.3.0
langchain-core>=0.3.0
langchain-community>=0.3.0  # Needed for tools, utilities, and older integrations

# 🔹 OpenAI Integration
langchain-openai>=0.2.0
openai>=1.52.0

# 🔹 Anthropic Integration
langchain-anthropic>=0.2.0
anthropic>=0.34.0

# 🔹 Google Gemini (PaLM) Integration
langchain-google-genai>=2.0.0
google-generativeai>=0.8.2

# 🔹 Hugging Face Integration
langchain-huggingface>=0.1.0
transformers>=4.45.0
huggingface-hub>=0.25.0

# 🔹 Environment Variable Management
python-dotenv>=1.0.1

# 🔹 Machine Learning Utilities
numpy>=1.26.4
scikit-learn>=1.5.2
